requirements.txt
streamlit==1.22.0
pandas==1.5.3
matplotlib==3.7.1
seaborn==0.12.2
requests==2.28.2
openai==0.27.6
scikit-learn==1.2.2
numpy==1.24.3
faiss-cpu==1.7.4
sentence-transformers==2.2.2